dhartisetu-backend/requirements.txt
tensorflow-cpu
torch
fastapi
uvicorn
huggingface_hub
joblib
numpy
pillow
